# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscribeRequest {
  /** Subscriber's email address */
  email: string;
  /** Which form/page captured the email (e.g. home, tackle-box) */
  source?: string;
}

export interface SubscribeResponse {
  success: boolean;
  message: string;
  alreadySubscribed: boolean;
}

export interface ErrorResponse {
  error: string;
}