# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subscribers
    description: Email list subscriber operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /subscribe:
    post:
      operationId: subscribe
      tags: [subscribers]
      summary: Subscribe to the email list
      description: Captures an email address with explicit consent. Idempotent — re-submitting an existing email returns success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: Subscribed successfully (or already subscribed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        "400":
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Subscriber's email address
        source:
          type: string
          description: Which form/page captured the email (e.g. home, tackle-box)
          default: home
      required:
        - email
    SubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        alreadySubscribed:
          type: boolean
      required:
        - success
        - message
        - alreadySubscribed
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error