# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Captures an email address with explicit consent. Idempotent — re-submitting an existing email returns success.
 * @summary Subscribe to the email list
 */
export const subscribeBodySourceDefault = `home`;

export const SubscribeBody = zod.object({
  email: zod.string().email().describe("Subscriber's email address"),
  source: zod
    .string()
    .default(subscribeBodySourceDefault)
    .describe("Which form\/page captured the email (e.g. home, tackle-box)"),
});

export const SubscribeResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  alreadySubscribed: zod.boolean(),
});